requirements_linux.txt
# ShowUI Linux环境依赖
# 适用于阿里云 NVIDIA A10 GPU

# 核心依赖
torch>=2.0.0
torchvision
torchaudio
transformers>=4.40.0
accelerate>=0.20.0
peft>=0.10.0
datasets>=2.14.0

# 量化和优化
bitsandbytes>=0.43.0
deepspeed>=0.14.0

# 数据处理
pillow>=9.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0
scikit-learn>=1.3.0
nltk>=3.8.0

# 可视化和监控
matplotlib>=3.7.0
tensorboard>=2.13.0
wandb>=0.15.0

# 开发工具
jupyterlab>=4.0.0
tqdm>=4.65.0

# API调用
openai>=1.0.0

# 性能优化库 (可选)
# flash-attn  # 需要编译，可能安装失败
# liger-kernel  # 可选的性能优化
